cloudfunctions/generate-contract/requirements.txt
tencentcloud-sdk-python==3.0.1035 